+++ src/SPECK3D_OMP_D.h
@@ -18,7 +18,7 @@ class SPECK3D_OMP_D {
   auto use_bitstream(const void*, size_t) -> RTNType;
 
 #ifndef QZ_TERM
-  auto set_bpp(float) -> RTNType;
+  auto set_bpp(double) -> RTNType;
 #endif
 
   void set_num_threads(size_t);
@@ -39,7 +39,7 @@ class SPECK3D_OMP_D {
   size_t m_num_threads = 1;                   // number of theads to use in OpenMP sections
 
 #ifndef QZ_TERM
-  float m_bpp = 0.0;
+  double m_bpp = 0.0;
 #endif
 
   const size_t m_header_magic = 26;  // header size would be this number + num_chunks * 4
